Add all changes after local test OK

+++ unattended_installer/install_functions/installMain.sh
@@ -180,7 +180,7 @@ function main() {
                 ;;
             "-dw"|"--download-wazuh")
                 if [ -z "${2}" ]; then
-                    common_logger -e "Error on arguments. Probably missing <deb|rpm> after -wd|--download-wazuh"
+                    common_logger -e "Error on arguments. Probably missing <deb|rpm> after -dw|--download-wazuh"
                     getHelp
                     exit 1
                 fi
@@ -222,7 +222,7 @@ function main() {
 
 # -------------- Preliminary checks  --------------------------------
 
-    if [ -z "${configurations}" ] && [ -z "${AIO}" ]; then
+    if [ -z "${configurations}" ] && [ -z "${AIO}" ] && [ -z "${download}" ]; then
         checks_previousCertificate
     fi
     checks_arch
@@ -258,14 +258,14 @@ function main() {
         common_logger "Created ${tar_file}. Contains Wazuh cluster key, certificates, and passwords necessary for installation."
     fi
 
-    if [ -z "${configurations}" ]; then
+    if [ -z "${configurations}" ] && [ -z "${download}" ]; then
         installCommon_extractConfig
         cert_readConfig
         rm -f "${config_file}"
     fi
 
     # Distributed architecture: node names must be different
-    if [[ -z "${AIO}" && ( -n "${indexer}"  || -n "${dashboard}" || -n "${wazuh}" )]]; then
+    if [[ -z "${AIO}" && -z "${download}" && ( -n "${indexer}"  || -n "${dashboard}" || -n "${wazuh}" )]]; then
         checks_names
     fi
 
@@ -355,7 +355,7 @@ function main() {
 
 # -------------------------------------------------------------------
 
-    if [ -z "${configurations}" ]; then
+    if [ -z "${configurations}" ] && [ -z "${download}" ]; then
         installCommon_restoreWazuhrepo
     fi
